refactor(units): SI luminous intensity header (#4)
